refactor: my page에 popup 컴포넌트 추가
